requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
docutils==0.20.1
executing==1.2.0
fonttools==4.39.4
idna==3.4
ijson==3.2.1
importlib-metadata==6.6.0
ipython==8.14.0
jaraco.classes==3.2.3
jedi==0.18.2
keyring==23.13.1
kiwisolver==1.4.4
markdown-it-py==2.2.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
numpy==1.24.3
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkginfo==1.9.6
plotly==5.14.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.8
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
readme-renderer==37.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.4.1
seaborn==0.12.2
six==1.16.0
sseclient-py==1.7.2
stack-data==0.6.2
tenacity==8.2.2
tqdm==4.65.0
traitlets==5.9.0
twine==4.0.2
typing_extensions==4.6.3
urllib3==2.0.2
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
wget==3.2
zipp==3.15.0
